sam-bedrock/pyproject.toml
[tool.poetry]
name = "sam-bedrock"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
slack-bolt = "1.18.0"
python-dotenv = "1.0.0"
boto3 = "^1.34.60"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
